+++ modules/data_crud/models/__init__.py
@@ -2,5 +2,6 @@
 Data CRUD Module - Models
 """
 from .crud_view import DcCrudView
+from .page_layout import DcPageLayout
 
-__all__ = ['DcCrudView']
+__all__ = ['DcCrudView', 'DcPageLayout']